Introduce `apple_rules_lint`
Composed of the regular pieces we need for this to fit neatly into the
Federation. The key things is in `workspace.bzl`, where we provide two
methods:

 * `register_linters`, designed to be the primary way that people use
   `apple_rules_lint`.
 * `ruleset_register_linters`, designed to be called from the
   `lang_setup` functions of other rulesets.

Ultimately, both of these functions do the same thing, but the former
will fail the build if not called in the right place.